requirements.txt
# Azure Functions dependencies
azure-functions==1.21.3

#azurefunctions-extensions-http-fastapi
# Obtaining azurefunctions-extensions-http-fastapi 1.0.0b3 from github to get pydantic 2.10.1 support
git+https://github.com/Azure/azure-functions-python-extensions.git@dev#egg=azurefunctions-extensions-http-fastapi&subdirectory=azurefunctions-extensions-http-fastapi

#azurefunctions-extensions-bindings-blob
# Obtaining azurefunctions-extensions-http-fastapi 1.0.0b3 from github to get pydantic 2.10.1 support
git+https://github.com/Azure/azure-functions-python-extensions.git@dev#egg=azurefunctions-extensions-bindings-blob&subdirectory=azurefunctions-extensions-bindings-blob

# Azure SDK dependencies
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-mgmt-web==8.0.0
azure-mgmt-storage==22.1.1
azure-appconfiguration==1.7.1
azure-appconfiguration-provider==2.0.1
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
msgraph-sdk==1.26.0

# Data handling and validation dependencies
jsonschema
python-dotenv==1.1.0
requests==2.32.3
openpyxl==3.1.5
tabulate==0.9.0

# AI and NLP dependencies
tiktoken==0.9.0
langchain-text-splitters==0.3.7
openai==1.69.0

# Error handling
tenacity==9.0.0

# Web and async dependencies
aiohttp==3.11.14
webvtt-py==0.5.1

# Image library
Pillow==11.1.0

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
